refactor: add _styles folder scss파일만 따로 관리
