+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import Header from './_components/Header';
 import Footer from './_components/Footer';
-import './globals.scss';
+import './_styles/globals.scss';
 
 const inter = Inter({ subsets: ['latin'] });
